isAbecedarian
Exercise 9.5 A word is said to be \abecedarian" if the letters in the word appear in alphabetical order. For example, the following are all six-letter. English abecedarian words:
abdest, acknow, acorsy, adempt, adipsy, agnosy, best, behint, be-know, bijoux, biopsy, cestuy, chintz, deflux, dehors, dehort, deinos,
diluvy, dimpsy
Write a method called isAbecedarian that takes a String and returns a boolean indicating whether the word is abecedarian. Your method can be iterative or recursive.